QuakeBot C/S
Hey mate, possible unlisted bot - 

QuakeBot C/S
https://web.archive.org/web/19990427091206/http://www.coe.uncc.edu:80/~jfrorie/quakebot.htm

What is a QuakeBot?
A QuakeBot is a program that runs on a system, connects to a Quake server and plays like a human opponent. The idea is to create a opponent that provides challenging deathmatch play. Theoretically, the players in the game shouldn't know the difference, unless the Bot/Programmer wants them to.

I grabbed what ZIPs I could from Archive.org - missing 2 by the look of it.
Looks to be source only - Haven't had a chance to compile and test, but it seems pretty interesting.
Let me know how it looks 🤔